last commit

+++ include/ast.h
@@ -2,13 +2,14 @@
 #define AST_H
 
 #include "lexer.h"
+#include "scope.h"
 
 typedef struct ASTNode {
-    TokenType type;
-    struct ASTNode *left;
-    struct ASTNode *right;
-    double value;
-    char var_name;
+    TokenType type;          // Type du nœud (e.g., token_while, token_int)
+    struct ASTNode* left;    // Sous-nœud gauche (e.g., condition pour while)
+    struct ASTNode* right;   // Sous-nœud droit (e.g., corps de la boucle)
+    int value;               // Valeur pour les nœuds contenant des entiers
+    char var_name;           // Nom de la variable pour les nœuds de type variable
 } ASTNode;
 
 ASTNode* create_node(TokenType type, ASTNode* left, ASTNode* right, int value, char var_name);
